requirements.txt
discord.py

cryptography

pynput
python-dotenv